merges AbstractAssetsMiddleware ans AssetsMiddleware
